png-renderer-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
playwright==1.42.0
pillow==10.2.0
pydantic==2.6.0
